+++ edgy/core/db/models/mixins/__init__.py
@@ -1,7 +1,6 @@
-from .date_parser import DateParser
 from .generics import DeclarativeMixin
 from .model_parser import ModelParser
 from .reflection import ReflectedModelMixin
 from .row import ModelRowMixin
 
-__all__ = ["DeclarativeMixin", "ModelRowMixin", "ReflectedModelMixin", "DateParser", "ModelParser"]
+__all__ = ["DeclarativeMixin", "ModelRowMixin", "ReflectedModelMixin", "ModelParser"]