+++ edgy/core/db/models/mixins/reflection.py
@@ -11,6 +11,7 @@ import sqlalchemy
 from pydantic_core._pydantic_core import SchemaValidator as SchemaValidator
 from sqlalchemy.ext.asyncio import AsyncConnection
 
+from edgy.core.connection.database import Database
 from edgy.core.utils.sync import run_sync
 from edgy.exceptions import ImproperlyConfigured
 
@@ -78,9 +79,9 @@ class ReflectedModelMixin:
                 raise e
 
         try:
-            async with registry.engine.begin() as connection:
-                table = await connection.run_sync(execute_reflection)
-            await registry.engine.dispose()
-            return table
+            async with Database(
+                registry.database, force_rollback=False
+            ) as database, database.transaction():
+                return await database.run_sync(execute_reflection)  # type: ignore
         except Exception as e:
             raise ImproperlyConfigured(detail=str(e)) from e