+++ edgy/core/utils/sync.py
@@ -1,6 +1,4 @@
 import asyncio
-from concurrent import futures
-from concurrent.futures import Future
 from typing import Any, Awaitable
 
 
@@ -9,8 +7,10 @@ def run_sync(async_function: Awaitable) -> Any:
     Runs the queries in sync mode
     """
     try:
-        return asyncio.run(async_function)
+        loop = asyncio.get_running_loop()
     except RuntimeError:
-        with futures.ThreadPoolExecutor(max_workers=1) as executor:
-            future: Future = executor.submit(asyncio.run, async_function)
-            return future.result()
+        loop = None
+    if loop:
+        return loop.run_until_complete(async_function)
+    else:
+        return asyncio.run(async_function)